# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: 云裁 MES / WIP API
  version: 1.0.0
  summary: 服装制造“一扎一码”生产管理接口契约
  description: |
    面向管理后台、工人扫码端、打印服务和外部集成的 REST API。

    核心约束：
    - 所有工厂业务请求必须携带 `X-Factory-Id`，服务端同时执行 RBAC 与数据范围校验。
    - 所有写请求必须携带 `Idempotency-Key`；相同键和相同请求体返回首次结果，相同键但不同请求体返回 `409 IDEMPOTENCY_KEY_REUSED`。
    - 生产事实时间以服务器时间为准，客户端时间仅用于弱网和离线审计。
    - 完工数量必须满足 `inputQty = goodQty + defectQty + missingQty`。
    - 完工、步骤更新、下一步骤解锁、计件快照、事件和审计在同一事务内提交。
    - 已完成生产事实不可覆盖；更正、返工和计件调整均以追加记录实现。
    - 金额以十进制定点字符串传输，禁止使用 IEEE 754 浮点数。
  contact:
    name: MES Product & Engineering
  license:
    name: Proprietary
servers:
  - url: https://mes.example.com
    description: 生产环境示例
  - url: http://127.0.0.1:3000
    description: 本地开发
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/workshops:
    get:
      tags: [Organization]
      operationId: listWorkshops
      summary: 查询车间
      x-permissions: [workshop:read]
      parameters:
        - $ref: '#/components/parameters/FactoryId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Status'
      responses:
        '200':
          description: 车间列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkshopPage' }
    post:
      tags: [Organization]
      operationId: createWorkshop
      summary: 创建车间
      x-permissions: [workshop:write]
      parameters:
        - $ref: '#/components/parameters/FactoryId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkshopUpsert' }
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workshop' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/production-lines:
    get:
      tags: [Organization]
      operationId: listProductionLines
      summary: 查询生产线
      x-permissions: [line:read]
      parameters:
        - $ref: '#/components/parameters/FactoryId'
        - $ref: '#/components/parameters/WorkshopIdQuery'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Status'
      responses:
        '200':
          description: 生产线列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductionLinePage' }
    post:
      tags: [Organization]
      operationId: createProductionLine
      summary: 创建生产线
      x-permissions: [line:write]
      parameters:
        - $ref: '#/components/parameters/FactoryId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductionLineUpsert' }
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductionLine' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/users:
    get:
      tags: [Access Control]
      operationId: listWorkerAccounts
      summary: 查询员工账号绑定状态
      x-permissions: [user:read]
      parameters:
        - $ref: '#/components/parameters/FactoryId'
      responses:
        '200':
          description: 员工账号列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkerAccountPage' }
    post:
      tags: [Access Control]
      operationId: createWorkerAccount
      summary: 为员工创建固定最小权限账号
      x-permissions: [user:write]
      parameters:
        - $ref: '#/components/parameters/FactoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkerAccountCreate' }
      responses:
        '201':
          description: 创建并绑定成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkerAccount' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /api/v1/users/{userId}/status:
    patch:
      tags: [Access Control]
      operationId: setWorkerAccountStatus
      summary: 启用或停用员工账号
      x-permissions: [user:write]
      parameters:
        - $ref: '#/components/parameters/FactoryId'
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkerAccountStatusPatch' }
      responses:
        '200':
          description: 状态更新成功，已有会话已撤销
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkerAccount' }
  /api/v1/users/{userId}/password:reset:
    post:
      tags: [Access Control]
      operationId: resetWorkerAccountPassword
      summary: 重置员工账号临时密码
      x-permissions: [user:write]
      parameters:
        - $ref: '#/components/parameters/FactoryId'
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkerAccountPasswordReset' }
      responses:
        '200':
          description: 密码已重置，已有会话已撤销
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkerAccount' }
...
  /api/v1/workers:
    get:
      tags: [Workers]
      operationId: listWorkers
      summary: 查询工人
      x-permissions: [worker:read]
      parameters:
        - $ref: '#/components/parameters/FactoryId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Status'
        - $ref: '#/components/parameters/WorkshopIdQuery'
        - $ref: '#/components/parameters/LineIdQuery'
        - name: processId
          in: query
          schema: { $ref: '#/components/schemas/Uuid' }
      responses:
        '200':
          description: 工人列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkerPage' }
    post:
      tags: [Workers]
      operationId: createWorker
      summary: 创建工人
      x-permissions: [worker:write]
      parameters:
        - $ref: '#/components/parameters/FactoryId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkerCreate' }
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Worker' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /api/v1/workers/{workerId}:
    parameters:
      - $ref: '#/components/parameters/FactoryId'
      - $ref: '#/components/parameters/WorkerId'
    get:
      tags: [Workers]
      operationId: getWorker
      summary: 获取工人及技能
      x-permissions: [worker:read]
      responses:
        '200':
          description: 工人详情
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Worker' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Workers]
      operationId: updateWorker
      summary: 更新工人
      x-permissions: [worker:write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkerPatch' }
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Worker' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/workers/{workerId}/skills:
    parameters:
      - $ref: '#/components/parameters/FactoryId'
      - $ref: '#/components/parameters/WorkerId'
    get:
      tags: [Workers]
      operationId: listWorkerSkills
      summary: 查询工人掌握的工种
      x-permissions: [worker:read]
      responses:
        '200':
          description: 技能列表
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/WorkerSkill' }
    put:
      tags: [Workers]
      operationId: replaceWorkerSkills
      summary: 替换工人技能集合
      description: 支持一个工人绑定多个工序；历史报工不受技能变更影响。
      x-permissions: [worker:skill:write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [skills]
              properties:
                skills:
                  type: array
                  minItems: 0
                  uniqueItems: true
                  items: { $ref: '#/components/schemas/WorkerSkillInput' }
      responses:
        '200':
          description: 技能已更新
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/WorkerSkill' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /api/v1/master-data/{resource}:
    parameters:
      - $ref: '#/components/parameters/FactoryId'
      - $ref: '#/components/parameters/MasterResource'
    get:
      tags: [Master Data]
      operationId: listMasterData
      summary: 查询基础资料
      description: 统一查询客户、款式、颜色、尺码和工序。
      x-permissions: [master:read]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Status'
      responses:
        '200':
          description: 基础资料列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MasterDataPage' }
    post:
      tags: [Master Data]
      operationId: createMasterData
      summary: 创建基础资料
      x-permissions: [master:write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MasterDataCreate' }
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MasterDataItem' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /api/v1/master-data/{resource}/{resourceId}:
    parameters:
      - $ref: '#/components/parameters/FactoryId'
      - $ref: '#/components/parameters/MasterResource'
      - $ref: '#/components/parameters/ResourceId'
    get:
      tags: [Master Data]
      operationId: getMasterData
      summary: 获取基础资料详情
      x-permissions: [master:read]
      responses:
        '200':
          description: 基础资料详情
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MasterDataItem' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Master Data]
      operationId: updateMasterData
      summary: 更新基础资料
      x-permissions: [master:write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MasterDataPatch' }
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MasterDataItem' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/processes/{processId}/adjust-rate:
    post:
      tags: [Master Data, Piecework]
      operationId: adjustProcessRate
      summary: 调整标准工序单价
      description: 更新工序默认价和未来路线；可选同步历史扎包。已完成报工保留原始金额，并生成正负计件差额记录。
      x-permissions: [piecework:adjust]
      parameters:
        - $ref: '#/components/parameters/FactoryId'
        - name: processId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Uuid' }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProcessRateAdjustmentRequest' }
      responses:
        '200':
          description: 工序单价已调整
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProcessRateAdjustmentResult' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 短期访问令牌；账号停用、设备解绑和会话撤销应立即失效。
  parameters:
    FactoryId:
      name: X-Factory-Id
      in: header
      required: true
      description: 当前业务工厂。服务端必须校验其属于令牌允许的数据范围。
      schema: { $ref: '#/components/schemas/Uuid' }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: UUID。相同键和请求体返回首次响应；相同键配不同请求体返回 409。
      schema: { type: string, format: uuid }
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: 资源乐观锁版本，格式为带引号的整数 ETag，例如 `"3"`。
      schema: { type: string, pattern: '^"[1-9][0-9]*"$' }
    Cursor:
      name: cursor
      in: query
      schema: { type: string, minLength: 1 }
      description: 上一页返回的不透明游标。
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Search:
      name: q
      in: query
      schema: { type: string, maxLength: 100 }
      description: 编码、名称或业务号模糊查询。
    Status:
      name: status
      in: query
      schema: { $ref: '#/components/schemas/ActiveStatus' }
    WorkshopIdQuery:
      name: workshopId
      in: query
      schema: { $ref: '#/components/schemas/Uuid' }
    LineIdQuery:
      name: productionLineId
      in: query
      schema: { $ref: '#/components/schemas/Uuid' }
    UserId:
      name: userId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
    WorkerId:
      name: workerId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
    ResourceId:
      name: resourceId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
    MasterResource:
      name: resource
      in: path
      required: true
      schema:
        type: string
        enum: [customers, styles, colors, sizes, processes]
...
  headers:
    XRequestId:
      description: 服务端请求追踪 UUID。
      schema: { type: string, format: uuid }
...
  responses:
...
    NotFound:
      description: 资源不存在；跨工厂资源也统一返回 404，避免信息泄露
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: 状态冲突、乐观锁失败、唯一约束冲突或幂等键误用
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationFailed:
      description: 字段、数量守恒或业务校验失败
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://mes.example.com/problems/quantity-unbalanced
            title: 完工数量不平
            status: 422
            code: QUANTITY_UNBALANCED
            detail: 投入 10，良品、次品和短缺合计 9，还差 1 件。
            requestId: a99a99c1-89dd-4dac-b87d-77d8ea9d74d1
            serverTime: '2026-08-29T11:00:00Z'
            errors:
              - field: missingQty
                code: SUM_MISMATCH
                message: 请补录 1 件。
...
  schemas:
    Uuid:
      type: string
      format: uuid
...
    ActiveStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
    UserStatus:
      type: string
      enum: [ACTIVE, LOCKED, INACTIVE]
    WorkerStatus:
      type: string
      enum: [ACTIVE, INACTIVE, LEFT]
...
    PageInfo:
      type: object
      required: [hasMore]
      properties:
        nextCursor: { type: [string, 'null'] }
        hasMore: { type: boolean }
        total:
          type: [integer, 'null']
          minimum: 0
          description: 仅在查询成本可接受时返回。
...
    WorkshopUpsert:
      type: object
      required: [code, name]
      properties:
        code: { type: string, minLength: 1, maxLength: 40 }
        name: { type: string, minLength: 1, maxLength: 120 }
        managerWorkerId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
        status: { $ref: '#/components/schemas/ActiveStatus' }
    Workshop:
      allOf:
        - type: object
          required: [id, factoryId, code, name, status]
          properties:
            id: { $ref: '#/components/schemas/Uuid' }
            factoryId: { $ref: '#/components/schemas/Uuid' }
            code: { type: string }
            name: { type: string }
            managerWorkerId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
            status: { $ref: '#/components/schemas/ActiveStatus' }
        - $ref: '#/components/schemas/AuditFields'
    WorkshopPage:
      type: object
      required: [items, page]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Workshop' } }
        page: { $ref: '#/components/schemas/PageInfo' }
    ProductionLineUpsert:
      type: object
      required: [workshopId, code, name]
      properties:
        workshopId: { $ref: '#/components/schemas/Uuid' }
        code: { type: string, minLength: 1, maxLength: 40 }
        name: { type: string, minLength: 1, maxLength: 120 }
        managerWorkerId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
        status: { $ref: '#/components/schemas/ActiveStatus' }
    ProductionLine:
      allOf:
        - type: object
          required: [id, factoryId, workshopId, code, name, status]
          properties:
            id: { $ref: '#/components/schemas/Uuid' }
            factoryId: { $ref: '#/components/schemas/Uuid' }
            workshopId: { $ref: '#/components/schemas/Uuid' }
            code: { type: string }
            name: { type: string }
            managerWorkerId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
            status: { $ref: '#/components/schemas/ActiveStatus' }
        - $ref: '#/components/schemas/AuditFields'
    ProductionLinePage:
      type: object
      required: [items, page]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/ProductionLine' } }
        page: { $ref: '#/components/schemas/PageInfo' }
...
    WorkerAccountCreate:
      type: object
      required: [workerId, username, displayName, password]
      additionalProperties: false
      properties:
        workerId: { $ref: '#/components/schemas/Uuid' }
        username: { type: string, minLength: 1, maxLength: 80 }
        displayName: { type: string, minLength: 1, maxLength: 100 }
        password: { type: string, minLength: 8, maxLength: 200, writeOnly: true }
    WorkerAccountStatusPatch:
      type: object
      required: [status]
      additionalProperties: false
      properties:
        status: { type: string, enum: [ACTIVE, INACTIVE] }
    WorkerAccountPasswordReset:
      type: object
      required: [password]
      additionalProperties: false
      properties:
        password: { type: string, minLength: 8, maxLength: 200, writeOnly: true }
    WorkerAccount:
      type: object
      required: [workerId, workerNo, workerName, workerStatus]
      properties:
        workerId: { $ref: '#/components/schemas/Uuid' }
        workerNo: { type: string }
        workerName: { type: string }
        workerStatus: { $ref: '#/components/schemas/WorkerStatus' }
        userId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
        username: { type: [string, 'null'] }
        displayName: { type: [string, 'null'] }
        accountStatus: { oneOf: [{ $ref: '#/components/schemas/UserStatus' }, { type: 'null' }] }
        lastLoginAt: { type: [string, 'null'], format: date-time }
    WorkerAccountPage:
      type: object
      required: [items, page]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/WorkerAccount' } }
        page: { $ref: '#/components/schemas/PageInfo' }
...
    WorkerCreate:
      type: object
      required: [workerNo, name]
      properties:
        workerNo: { type: string, minLength: 1, maxLength: 40 }
        name: { type: string, minLength: 1, maxLength: 80 }
        userId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
        pin: { type: [string, 'null'], minLength: 4, maxLength: 20, writeOnly: true }
        workshopId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
        productionLineId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
        hiredOn: { type: [string, 'null'], format: date }
        status: { $ref: '#/components/schemas/WorkerStatus' }
        skills:
          type: array
          items: { $ref: '#/components/schemas/WorkerSkillInput' }
    WorkerPatch:
      type: object
      minProperties: 1
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        userId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
        pin: { type: [string, 'null'], minLength: 4, maxLength: 20, writeOnly: true }
        workshopId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
        productionLineId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
        status: { $ref: '#/components/schemas/WorkerStatus' }
        leftOn: { type: [string, 'null'], format: date }
    Worker:
      allOf:
        - type: object
          required: [id, factoryId, workerNo, name, status, skills]
          properties:
            id: { $ref: '#/components/schemas/Uuid' }
            factoryId: { $ref: '#/components/schemas/Uuid' }
            userId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
            workerNo: { type: string }
            name: { type: string }
            workshopId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
            productionLineId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
            status: { $ref: '#/components/schemas/WorkerStatus' }
            hiredOn: { type: [string, 'null'], format: date }
            leftOn: { type: [string, 'null'], format: date }
            skills:
              type: array
              items: { $ref: '#/components/schemas/WorkerSkill' }
        - $ref: '#/components/schemas/AuditFields'
    WorkerPage:
      type: object
      required: [items, page]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Worker' } }
        page: { $ref: '#/components/schemas/PageInfo' }
    WorkerSkillInput:
      type: object
      required: [processId, level]
      properties:
        processId: { $ref: '#/components/schemas/Uuid' }
        level: { type: integer, minimum: 1, maximum: 5 }
        effectiveFrom: { type: [string, 'null'], format: date }
        effectiveTo: { type: [string, 'null'], format: date }
    WorkerSkill:
      allOf:
        - $ref: '#/components/schemas/WorkerSkillInput'
        - type: object
          required: [id, workerId, processCode, processName]
          properties:
            id: { $ref: '#/components/schemas/Uuid' }
            workerId: { $ref: '#/components/schemas/Uuid' }
            processCode: { type: string }
            processName: { type: string }
    MasterDataCreate:
      type: object
      required: [code, name]
      properties:
        code: { type: string, minLength: 1, maxLength: 60 }
        name: { type: string, minLength: 1, maxLength: 120 }
        status: { $ref: '#/components/schemas/ActiveStatus' }
        customerId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
        customerStyleNo: { type: [string, 'null'], maxLength: 100 }
        imageUrl: { type: [string, 'null'], format: uri }
        versionName: { type: [string, 'null'], maxLength: 40 }
        displayOrder: { type: integer, minimum: 0, default: 0 }
        unit: { type: [string, 'null'], maxLength: 30 }
        defaultStandardSeconds: { type: [integer, 'null'], minimum: 0 }
        defaultPieceRate: { oneOf: [{ $ref: '#/components/schemas/NonNegativeMoney' }, { type: 'null' }] }
    MasterDataPatch:
      type: object
      minProperties: 1
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        status: { $ref: '#/components/schemas/ActiveStatus' }
        customerId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
        customerStyleNo: { type: [string, 'null'], maxLength: 100 }
        imageUrl: { type: [string, 'null'], format: uri }
        versionName: { type: [string, 'null'], maxLength: 40 }
        displayOrder: { type: integer, minimum: 0 }
        unit: { type: [string, 'null'], maxLength: 30 }
        defaultStandardSeconds: { type: [integer, 'null'], minimum: 0 }
        defaultPieceRate: { oneOf: [{ $ref: '#/components/schemas/NonNegativeMoney' }, { type: 'null' }] }
    MasterDataItem:
      allOf:
        - type: object
          required: [id, factoryId, resourceType, code, name, status]
          properties:
            id: { $ref: '#/components/schemas/Uuid' }
            factoryId: { $ref: '#/components/schemas/Uuid' }
            resourceType: { type: string, enum: [customers, styles, colors, sizes, processes] }
            code: { type: string }
            name: { type: string }
            status: { $ref: '#/components/schemas/ActiveStatus' }
            customerId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
            customerStyleNo: { type: [string, 'null'] }
            imageUrl: { type: [string, 'null'], format: uri }
            versionName: { type: [string, 'null'] }
            displayOrder: { type: [integer, 'null'] }
            unit: { type: [string, 'null'] }
            defaultStandardSeconds: { type: [integer, 'null'] }
            defaultPieceRate: { oneOf: [{ $ref: '#/components/schemas/NonNegativeMoney' }, { type: 'null' }] }
        - $ref: '#/components/schemas/AuditFields'
    MasterDataPage:
      type: object
      required: [items, page]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/MasterDataItem' } }
        page: { $ref: '#/components/schemas/PageInfo' }
    ProcessRateAdjustmentRequest:
      type: object
      required: [expectedVersion, unitRate, applyToHistoricalBundles]
      properties:
        expectedVersion: { type: integer, minimum: 1 }
        unitRate: { $ref: '#/components/schemas/NonNegativeMoney' }
        applyToHistoricalBundles: { type: boolean, default: false }
        reason:
          type: [string, 'null']
          minLength: 1
          maxLength: 500
          description: 调整历史扎包时必填。
...